move logger macros into core project
