+++ core/shared/src/main/scala/scribe/Loggable.scala
